fix(pix3d): int32-wrap the GouraudRaster DivTable product (B-01 parity backport)
GouraudRaster's colour-interpolation step `var8 = ((arg7-arg6)*DivTable[arg3])
>> 15` is a 32-bit int product in the pinned Client-Java (176a85f7,
Pix3D.java:868) that wraps mod 2^32 and overflows on steep colour deltas over
short spans. rev-254 wraps it (audit pix3d-B-01); rev-245.2 did not. Land the
same int(int32(...)) wrap for Java parity — the same masked-but-real precedent
as the F2 C-01 port.

Masked for all current goldens/icons: overflow needs |arg7-arg6| >= 2^31/32768
= 65_536, far above any golden colour delta. Proof of no output change: all tri
goldens + palette stay GREEN, and the full 2978-icon icondump run is
byte-identical before and after (hash 6eba4fe1…).

(The TextureRaster shade analogue pix3d-3-03 was already wrapped here by
0700e33 audit G3; this closes the remaining DivTable-product gap on 245.2.)

+++ pkg/jagex2/graphics/pix3d/pix3d.go
@@ -910,7 +910,12 @@ func GouraudRaster(arg0 []int, arg1, arg4, arg5, arg6, arg7 int) {
 			arg1 += arg4
 			arg3 = (arg5 - arg4) >> 2
 			if arg3 > 0 {
-				var8 = ((arg7 - arg6) * DivTable[arg3]) >> 15
+				// Java: Pix3D.java:868 (pin 176a85f7) — 32-bit product wraps before
+				// the >>15 (overflows on steep colour deltas over short spans; audit
+				// pix3d-B-01, parity backport of rev-254's wrap). Masked for the tri
+				// goldens/icons: overflow needs |arg7-arg6| >= 2^31/32768 = 65_536,
+				// far above any golden colour delta.
+				var8 = int(int32((arg7-arg6)*DivTable[arg3])) >> 15
 			} else {
 				var8 = 0
 			}
